+++ dapr/clients/__init__.py
@@ -8,13 +8,11 @@ Licensed under the MIT License.
 from dapr.clients.base import DaprActorClientBase
 from dapr.clients.exceptions import DaprInternalError, ERROR_CODE_UNKNOWN
 from dapr.clients.http.dapr_actor_http_client import DaprActorHttpClient
-from dapr.clients.grpc.dapr_client import (
-    MetadataDict,
-    MetadataTuple,
-    DaprClient,
-    InvokeServiceRequestData,
-    InvokeServiceResponse,
-)
+
+from dapr.clients.grpc.client_request import InvokeServiceRequestData
+from dapr.clients.grpc.client_response import InvokeServiceResponse
+from dapr.clients.grpc.client import DaprClient
+
 
 __all__ = [
     'DaprActorClientBase',
@@ -22,8 +20,6 @@ __all__ = [
     'DaprInternalError',
     'ERROR_CODE_UNKNOWN',
     'DaprClient',
-    'MetadataTuple',
-    'MetadataDict',
     'InvokeServiceRequestData',
     'InvokeServiceResponse',
 ]
