+++ dapr/clients/grpc/_helpers.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+"""
+Copyright (c) Microsoft Corporation.
+Licensed under the MIT License.
+"""
+
+from typing import Dict, List, Union, Tuple
+
+MetadataDict = Dict[str, List[Union[bytes, str]]]
+MetadataTuple = Tuple[Tuple[str, Union[bytes, str]], ...]
